requirements.txt
python==3.7
transformers==4.2.2
pytorch==1.7.1
amrlib==0.5.0
nltk==3.5
scipy==1.6.3
